pyproject.toml
[tool.pdm]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Amanda de Oliveira Pires", email = "nanapires255@gmail.com"},
]
dependencies = [
    "django>=4.2.7",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.2.0",
    "djangorestframework-simplejwt>=5.2.2",
    "setuptools>=68.1.0",
    "python-magic>=0.4.27",
    "Pillow>=10.0.0",
    "django-filter>=23.2",
    "netifaces>=0.11.0",
    "python-dotenv>=1.0.0",
    "gunicorn>=21.2.0",
    "whitenoise[brotli]>=6.6.0",
    "psycopg2-binary>=2.9.9",
    "django-cloudinary-storage>=0.3.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"


[tool.pdm.scripts]
pre_dev = "python ./scripts/set_my_ip.py"
dev = "python manage.py runserver 0.0.0.0:19003"